Extracted URI builder logic, fixed regex expressions and added unit tests.
